Update BoatsCard test and remove fixture duplication
