+++ src/domain/individualCustomer/boatsCard/__fixtures__/mockData.ts
@@ -1,4 +1,4 @@
-import { BoatCert } from '../../types';
+import { Boat, BoatCert, LargeBoat } from '../../types';
 import { BoatCertificateType } from '../../../../@types/__generated__/globalTypes';
 
 export const insuraceCert: BoatCert = {
@@ -16,4 +16,38 @@ export const inspectionCert: BoatCert = {
   validUntil: '2099-09-04',
   checkedAt: '2020-04-28',
   checkedBy: 'Jukka Virtanen',
+};
+
+export const boat: Boat = {
+  id: '1111',
+  boatType: {
+    id: '7',
+    name: 'Troolari',
+  },
+  width: 2,
+  length: 10,
+  draught: 2,
+  weight: 15000,
+  name: 'Huh hah hei',
+  model: 'Trawl 11',
+  registrationNumber: 'TRLR-555',
+};
+
+export const largeBoat: LargeBoat = {
+  id: '9999',
+  boatType: {
+    id: '8',
+    name: 'Suuri alus (yli 20t)',
+  },
+  width: 4,
+  length: 15,
+  draught: 2,
+  weight: 15000,
+  name: 'DevilsDriver',
+  model: 'Trawl 77',
+  registrationNumber: 'TRLR-666',
+  propulsion: 'Bensin',
+  hullMaterial: 'Aluminium',
+  intendedUse: 'For fun',
+  certificates: [inspectionCert, insuraceCert],
 };
